+++ test/PHPStan/data/template-variables/templates/rows.html.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+use SubstancePHP\HTTP\Renderer\HtmlRenderer;
+
+/** @var HtmlRenderer $this */
+/** @var array<string, string> $row */
+?>
+<div>
+    <?= $this->partial('detail', $row) ?>
+</div>